Combine RenderingResourcesSystem and SpriteBatchSystem
