+++ Features/Sprites/Systems/SpriteRenderSystem.cs
@@ -3,7 +3,6 @@ using Arch.System;
 using Arch.System.SourceGenerator;
 using Exanite.Ecs.Systems;
 using Exanite.GravitationalTetris.Features.Cameras.Components;
-using Exanite.GravitationalTetris.Features.Rendering;
 using Exanite.GravitationalTetris.Features.Sprites.Components;
 using Exanite.GravitationalTetris.Features.Transforms.Components;
